[POC] Add config option to purge classes from section, and routes from navigation configs
First naive proof-of-concept approach to remove parts of the appinfo.xml.

The removed nodes "classes" are no further differentated in config.php
because one class is not used in two config nodes at the same time.
This simplifies the config.php structure.

Nodes here refers to the XML tree nodes.