(C11N) .gitlab-ci.yml: set GIT_SUBMODULE_FORCE_HTTPS
Attempt to fix:

  Cloning into '/builds/strato-dev/nextcloud-server/custom-apps/enc_analytics'...
  remote: The project you were looking for could not be found or you don't have permission to view it.
  fatal: repository 'https://gitlab.git-moc.server.lan/strato-dev/enc_analytics.git/' not found

The logged location can indeed not be found. The "/" suffix seems to be
wrong. When manually requesting the URL without the trailing "/" it can
be opened.

Although GitLab appending the slash seems to be quite arbitrary, this
config was added as desperate attempt to make it work.

> For submodules located on the same GitLab server and configured with
> a Git or SSH URL, make sure you set the GIT_SUBMODULE_FORCE_HTTPS
> variable. [1]

[1]: https://docs.gitlab.com/ee/ci/git_submodules.html#use-git-submodules-in-cicd-jobs