[POC] lib: JSResourceLocator can ignore script paths in /app during lookup
The resource locator takes app names and path fragments and probes the
file system for the presence of a script file. Files are probed in an
order that prefers the /dist directory before app specific paths.

The resource locator can now skip these paths if an app was configured
to be forcefully loaded from another apps directory.

The config

   "forced_app_dirs" = ["<app_name>" => "<app dir name>"]

allows to force app manager to skip loading apps from "/apps" and
instread try loading it from another directory. It was introduced in a
previous commit.