[POC] lib: JSResourceLocator can ignore assets by config
The resource locator takes app names and path fragments and probes the
file system for the presence of a script file. Files are then included
into the page for loading by the browser.

The resource locator can now skip these paths if such a path was
configured to be disabled.

The intention of this config is to prevent frontend code of certain core
apps from being loaded so that a custom app can load their own scripts.

The config

   "disabled_scripts" = [
      "dist/foo.js",
      "apps/foo_app/js/bar.js",
   ]

would prevent the core script "dist/foo.js" and the "foo_app"s "bar.js"
to be loaded.