[DUMMY;DROP] Transplant older version of @volar/language-core for testing
This is just a test to see whether the fix of package-lock helped.

== Idea

Take an _older_ version pinning and write it into the package-lock to
ensure npm ci will keep it and not complain about conflicts.

== How?

Package-lock for that dependency inspected

   $ git show cd6bb242d427434985368eae36d99a952262473f:package-lock.json \
     | grep -A 20 '"node_modules/@volar/language-core"'

The object was copied and the object in package-lock.json replaced with
this one.

== Test/proof

   $ rm -rf node_modules/
   $ npm ci
   $ npm why @volar/language-core | head -n1
     @volar/language-core@2.4.18 dev

Success!