IONOS(user_oidc): update user_oidc submodule to v7.4.0
With the previous update to v7.3.2 we tried to fix a bug with the
logout. With tests of v7.3.2 on pre-live we discovered that logout
does no longer work at all.

The logout hits the /logout endpoint of Nextcloud, not the
/apps/user_oidc/sls endpoint of user_oidc.

Some of the changes to v7.4.0 [1] sound they may fix this.

We try this version first before debugging more.

[1]: https://github.com/nextcloud/user_oidc/blob/main/CHANGELOG.md#740--2025-09-01

Signed-off-by: Thomas Lehmann <t.lehmann@strato.de>